+++ src/DocsTestRunner.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace VacantPlanet\Development;
+namespace FiveOrbs\Development;
 
 class DocsTestRunner
 {