+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Change namespace to VacantPlanet
+- Change namespace to FiveOrbs
 
 ## [1.1.1] - 2023-11-24
 
@@ -63,4 +63,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- php-cs-fixer rules via `VacantPlanet\Development\PhpCsFixer\Config`.
+- php-cs-fixer rules via `FiveOrbs\Development\PhpCsFixer\Config`.